requirements.txt
alpha-vantage
streamlit
kafka-python
pyspark==3.5.0
pandas